Conflicts:
	app/socal/foro.py
	app/socal/ident.py
	static/app/foro/VComentariosPagina.html
	static/app/foro/VHilos.html
	static/app/ident/VPrincipal.html
	static/app/ident/VSecundaria.html
	static/js/app/foro/ForoController.js
	static/js/app/ident/IdentController.js
	static/js/app/ident/IdentService.js

+++ app/socal/foro.py
@@ -16,7 +16,6 @@ def VForo():
     res = {}
     if "actor" in session:
         res['actor']=session['actor']
-        res['usuario'] = {'nombre': session['nombre_usuario']}
     #Action code goes here, res should be a JSON structure
     
     listaHilos = [{
@@ -25,6 +24,9 @@ def VForo():
         'fecha': h.fecha_creacion,
         'autor': h.raiz.autor_id}
         for h in Hilo.query.filter_by(foro_id=idForo)]
+
+    if 'nombre_usuario' in session:
+        res['idUsuario'] = session['nombre_usuario']
 
     res['data'] = listaHilos
 
@@ -38,13 +40,15 @@ def VForos():
     res = {}
     if "actor" in session:
         res['actor']=session['actor']
-        res['usuario'] = {'nombre': session['nombre_usuario']}
     #Action code goes here, res should be a JSON structure
     listaForos = []
     for ftitulo, ffecha,fautor in db.session.query(Foro.titulo, Foro.fecha_creacion,Foro.autor_id):
         listaForos.append({'titulo':ftitulo,'fecha': ffecha,'autor':fautor})
 
-    res['data'] = listaForos
+    res['data'] = listaForos    
+
+    if 'nombre_usuario' in session:
+        res['idUsuario'] = session['nombre_usuario']
 
     #Action code ends here
     return json.dumps(res)
@@ -58,11 +62,15 @@ def AgregForo():
     {'label':'/VForos', 'msg':['No se pudo agregar el nuevo foro']}]
 
     titulo_nuevo_foro = params['texto']
-    nuevo_foro = Foro(titulo=titulo_nuevo_foro, nombre_usuario=session['nombre_usuario'])
-    db.session.add(nuevo_foro)
-    db.session.commit()
+    foro_ya_existe = Foro.query.filter_by(titulo=titulo_nuevo_foro).first()
+    res = results[1]
+    if foro_ya_existe is None:
+        if 'nombre_usuario' in session:
+            nuevo_foro = Foro(titulo=titulo_nuevo_foro, nombre_usuario=session['nombre_usuario'])
+            db.session.add(nuevo_foro)
+            db.session.commit()
+            res = results[0]
 
-    res = results[0]
     return json.dumps(res)
 
 #------------------------------------------------------------------------------#
@@ -76,7 +84,6 @@ def VHilos():
     idHilo = request.args['idHilo']
     if "actor" in session:
         res['actor']=session['actor']
-        res['usuario'] = {'nombre': session['nombre_usuario']}
     #Action code goes here, res should be a JSON structure
 
     hilo = Hilo.query.filter_by(id=idHilo).first()
@@ -85,6 +92,9 @@ def VHilos():
         res['foroPadre'] =  hilo.foro_id
     res['tituloNuevaPublicacion'] = "RE: " + raiz.titulo
     res['publicacion'] = raiz.a_diccionario()
+
+    if 'nombre_usuario' in session:
+        res['idUsuario'] = session['nombre_usuario']
 
     #Action code ends here
     return json.dumps(res)
@@ -97,6 +107,8 @@ def AgregHilo():
     results = [{'label':'/VForo/'+session['idForo'], 'msg':['Hilo Agregado']}, 
     {'label':'/VForos/'+session['idForo'], 'msg':['No se pudo agregar el nuevo hilo']}]
 
+    if 'nombre_usuario' not in session:
+        return json.dump(results[1])
     titulo_publicacion  = params['titulo']
     contenido_publicacion = params['contenido']
 
@@ -126,6 +138,9 @@ def AElimForo():
     results = [{'label':'/VForos', 'msg':['Foro eliminado']},
      {'label':'/VForos', 'msg':['No se pudo eliminar el foro']}, ]
     
+    if 'nombre_usuario' not in session:
+        return json.dump(results[1])
+    
     foro_a_eliminar = Foro.query.filter_by(titulo=idForo).first()
     if foro_a_eliminar.autor_id == session['nombre_usuario']:
         db.session.delete(foro_a_eliminar)
@@ -145,6 +160,9 @@ def AElimHilo():
     results = [{'label':'/VForo/'+session['idForo'], 'msg':['Hilo eliminado']},
      {'label':'/VForo/'+session['idForo'], 'msg':['No se pudo eliminar el hilo']}, ]
     
+    if 'nombre_usuario' not in session:
+        return json.dump(results[1])
+        
     hilo_a_eliminar = Hilo.query.filter_by(id=idHilo).first()
     if hilo_a_eliminar.raiz.autor_id == session['nombre_usuario']:
         db.session.delete(hilo_a_eliminar)
@@ -165,6 +183,9 @@ def AgregPublicacion():
     params = request.get_json()
     results = [{'msg':['Respuesta enviada']},{'msg':['No se pudo enviar la respuesta']}]
     
+    if 'nombre_usuario' not in session:
+        return json.dump(results[1])
+        
     idPadre = params['id']
     titulo = params['titulo']
     contenido = params['contenido']
@@ -193,6 +214,9 @@ def AElimPublicacion():
     publicacion_a_eliminar = Publicacion.query.filter_by(id=idPublicacion).first()
     idHilo = publicacion_a_eliminar.hilo_id
     results = [{'msg':['Publicacion eliminada']}, {'msg':['No se pudo eliminar la publicacion']}]
+    
+    if 'nombre_usuario' not in session:
+        return json.dump(results[1])
     
     if publicacion_a_eliminar.autor_id == session['nombre_usuario']:
         if publicacion_a_eliminar.hijos == []:
